Add a shadow to devices
This is necessary especially for the BBC micro:bit, as it is nearly
black and would otherwise be hard to distinguish from the black
background.

There is also the drop-shadow CSS property but unfortunately it doesn't
work yet in Chrome / VS Code. It works in Firefox however. See this
Chrome bug: https://bugs.chromium.org/p/chromium/issues/detail?id=109224