Do not depend on the Go extension
This may seem strange, but not depending on this version makes it easier
to fix issues after running `tinygo clean` when the Go extension fails
to startup. Just select a board again in TinyGo, restart VSCode, and
things work again.

TODO: the extension should also detect this situation and resolve it, so
that there's no need for manually selecting a different board.